test(generation): split tests in multiple files
