--- tests/inference/test_modeling_decoder.py
@@ -1,172 +1,0 @@
-# coding=utf-8
-# Copyright 2022 The HuggingFace Team. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-import os
-from tempfile import TemporaryDirectory
-
-import pytest
-import torch
-from huggingface_hub import HfApi
-from transformers import AutoTokenizer
-from transformers.testing_utils import ENDPOINT_STAGING
-
-from optimum.neuron import NeuronModelForCausalLM
-from optimum.neuron.utils.testing_utils import is_inferentia_test, requires_neuronx
-from optimum.utils import logging
-from optimum.utils.testing_utils import TOKEN, USER
-
-from .inference_utils import MODEL_NAMES
-
-
-logger = logging.get_logger()
-
-
-DECODER_MODEL_ARCHITECTURES = ["gpt2"]
-
-
-@pytest.fixture(scope="module", params=[MODEL_NAMES[model_arch] for model_arch in DECODER_MODEL_ARCHITECTURES])
-def export_model_id(request):
-    return request.param
-
-
-@pytest.fixture(scope="module")
-@requires_neuronx
-def neuron_model_path(export_model_id):
-    # For now we need to use a batch_size of 2 because it fails with batch_size == 1
-    model = NeuronModelForCausalLM.from_pretrained(export_model_id, export=True, batch_size=2)
-    model_dir = TemporaryDirectory()
-    model_path = model_dir.name
-    model.save_pretrained(model_path)
-    del model
-    tokenizer = AutoTokenizer.from_pretrained(export_model_id)
-    tokenizer.save_pretrained(model_path)
-    del tokenizer
-    # Yield instead of returning to keep a reference to the temporary directory.
-    # It will go out of scope and be released only once all tests needing the fixture
-    # have been completed.
-    yield model_path
-
-
-@pytest.fixture(scope="module")
-def neuron_push_id(export_model_id):
-    model_name = export_model_id.split("/")[-1]
-    repo_id = f"{USER}/{model_name}-neuronx"
-    return repo_id
-
-
-def _check_neuron_model(neuron_model, batch_size=None, sequence_length=None, num_cores=None, auto_cast_type=None):
-    neuron_config = getattr(neuron_model.config, "neuron", None)
-    assert neuron_config
-    if batch_size:
-        assert neuron_config["batch_size"] == batch_size
-    if num_cores:
-        assert neuron_config["num_cores"] == num_cores
-    if auto_cast_type:
-        assert neuron_config["auto_cast_type"] == auto_cast_type
-
-
-@pytest.mark.parametrize(
-    "batch_size, sequence_length, num_cores, auto_cast_type",
-    [
-        [1, 128, 2, "fp32"],
-        [1, 128, 2, "fp32"],
-        [2, 512, 2, "bf16"],
-    ],
-)
-@is_inferentia_test
-@requires_neuronx
-def test_model_export(export_model_id, batch_size, sequence_length, num_cores, auto_cast_type):
-    model = NeuronModelForCausalLM.from_pretrained(
-        export_model_id,
-        export=True,
-        batch_size=batch_size,
-        sequence_length=sequence_length,
-        num_cores=num_cores,
-        auto_cast_type=auto_cast_type,
-    )
-    _check_neuron_model(model, batch_size, sequence_length, num_cores, auto_cast_type)
-
-
-@is_inferentia_test
-@requires_neuronx
-def test_model_from_path(neuron_model_path):
-    model = NeuronModelForCausalLM.from_pretrained(neuron_model_path)
-    _check_neuron_model(model)
-
-
-@is_inferentia_test
-@requires_neuronx
-def test_model_from_hub():
-    model = NeuronModelForCausalLM.from_pretrained(
-        "dacorvo/tiny-random-gpt2-neuronx", revision="b8f1aec89f9b278721068bfe616fa9227c1d0238"
-    )
-    _check_neuron_model(model, batch_size=16, num_cores=2, auto_cast_type="fp32")
-
-
-def _test_model_generation(model, tokenizer, batch_size, input_length, **gen_kwargs):
-    input_ids = torch.ones((batch_size, input_length), dtype=torch.int64)
-    with torch.inference_mode():
-        sample_output = model.generate(input_ids, **gen_kwargs)
-        assert sample_output.shape[0] == batch_size
-
-
-@pytest.mark.parametrize(
-    "gen_kwargs",
-    [
-        {"do_sample": True},
-        {"do_sample": True, "temperature": 0.7},
-        {"do_sample": False},
-        {"do_sample": False, "repetition_penalty": 1.2},
-    ],
-    ids=["sample", "sample-with-temp", "greedy", "greedy_no-repeat"],
-)
-@is_inferentia_test
-@requires_neuronx
-def test_model_generation(neuron_model_path, gen_kwargs):
-    model = NeuronModelForCausalLM.from_pretrained(neuron_model_path)
-    tokenizer = AutoTokenizer.from_pretrained(neuron_model_path)
-    _test_model_generation(model, tokenizer, model.batch_size, 10, **gen_kwargs)
-
-
-@is_inferentia_test
-@requires_neuronx
-def test_model_generation_input_dimensions(neuron_model_path):
-    model = NeuronModelForCausalLM.from_pretrained(neuron_model_path)
-    tokenizer = AutoTokenizer.from_pretrained(neuron_model_path)
-    # Using valid input dimensions
-    _test_model_generation(model, tokenizer, model.batch_size, model.max_length // 2)
-    # Using an incompatible batch_size
-    with pytest.raises(ValueError, match="The specified batch_size"):
-        _test_model_generation(model, tokenizer, model.batch_size + 1, model.max_length)
-    # Using an incompatible input length
-    with pytest.raises(ValueError, match="The input sequence length"):
-        _test_model_generation(model, tokenizer, model.batch_size, input_length=model.max_length * 2)
-
-
-@is_inferentia_test
-@requires_neuronx
-def test_push_to_hub(neuron_model_path, neuron_push_id):
-    model = NeuronModelForCausalLM.from_pretrained(neuron_model_path)
-    model.push_to_hub(neuron_model_path, neuron_push_id, use_auth_token=TOKEN, endpoint=ENDPOINT_STAGING)
-    api = HfApi(endpoint=ENDPOINT_STAGING, token=TOKEN)
-    try:
-        hub_files_info = api.list_files_info(neuron_push_id)
-        hub_files_path = [info.rfilename for info in hub_files_info]
-        for path, _, files in os.walk(neuron_model_path):
-            for name in files:
-                local_file_path = os.path.join(path, name)
-                hub_file_path = os.path.relpath(local_file_path, neuron_model_path)
-                assert hub_file_path in hub_files_path
-    finally:
-        api.delete_repo(neuron_push_id)
